+++ crates/fluvio-smartengine/src/engine/wasmtime/transforms/map.rs
@@ -10,22 +10,21 @@ mod test {
 
     use crate::engine::{
         SmartEngine, SmartModuleChainBuilder, SmartModuleConfig, metrics::SmartModuleChainMetrics,
-        wasmtime_engine::transforms::simple_transform::FILTER_MAP_FN_NAME,
+        wasmtime::transforms::simple_transform::MAP_FN_NAME,
     };
-
-    const SM_FILTER_MAP: &str = "fluvio_smartmodule_filter_map";
-
     use crate::engine::fixture::read_wasm_module;
+
+    const SM_MAP: &str = "fluvio_smartmodule_map";
 
     #[ignore]
     #[test]
-    fn test_filter_map() {
+    fn test_map() {
         let engine = SmartEngine::new();
         let mut chain_builder = SmartModuleChainBuilder::default();
 
         chain_builder.add_smart_module(
             SmartModuleConfig::builder().build().unwrap(),
-            read_wasm_module(SM_FILTER_MAP),
+            read_wasm_module(SM_MAP),
         );
 
         let mut chain = chain_builder
@@ -34,15 +33,16 @@ mod test {
 
         assert_eq!(
             chain.instances().first().expect("first").transform().name(),
-            FILTER_MAP_FN_NAME
+            MAP_FN_NAME
         );
 
         let metrics = SmartModuleChainMetrics::default();
-        let input = vec![Record::new("10"), Record::new("11")];
+        let input = vec![Record::new("apple"), Record::new("fruit")];
         let output = chain
             .process(SmartModuleInput::try_from(input).expect("input"), &metrics)
             .expect("process");
-        assert_eq!(output.successes.len(), 1); // one record passed
-        assert_eq!(output.successes[0].value.as_ref(), b"5");
+        assert_eq!(output.successes.len(), 2); // one record passed
+        assert_eq!(output.successes[0].value.as_ref(), b"APPLE");
+        assert_eq!(output.successes[1].value.as_ref(), b"FRUIT");
     }
 }
