+++ crates/fluvio-smartengine/src/engine/wasmtime/transforms/simple_transform.rs
@@ -4,7 +4,7 @@ use fluvio_smartmodule::dataplane::smartmodule::{
 use wasmtime::{TypedFunc, AsContextMut};
 use anyhow::Result;
 
-use crate::engine::wasmtime_engine::{
+use crate::engine::wasmtime::{
     instance::{SmartModuleInstanceContext, SmartModuleTransform},
     state::WasmState,
 };
